requirements.txt
pyyaml
requests
redis
